requirements.txt
# Core dependencies
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.2.0
xgboost>=1.7.0

# API and serving
fastapi>=0.95.0
uvicorn>=0.20.0
pydantic>=1.10.0
python-multipart>=0.0.5

# Dashboard and visualization  
streamlit>=1.25.0
matplotlib>=3.6.0
seaborn>=0.12.0
plotly>=5.11.0

# Data processing
cicflowmeter>=0.4.2
pyarrow>=10.0.0
joblib>=1.2.0

# Configuration and utilities
pyyaml>=6.0
python-dotenv>=0.19.0
click>=8.0.0

# Logging and monitoring
loguru>=0.6.0
prometheus-client>=0.15.0

# Testing (optional)
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0
mypy>=0.991

# Additional dependencies for Windows compatibility
psutil>=5.8.0
